BC-239 moved summary classes
