+++ src/main/java/uk/gov/justice/laa/amend/claim/viewmodels/CivilClaimSummary.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.amend.claim.models;
+package uk.gov.justice.laa.amend.claim.viewmodels;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
