+++ src/main/java/uk/gov/justice/laa/amend/claim/viewmodels/ClaimSummary.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.amend.claim.models;
+package uk.gov.justice.laa.amend.claim.viewmodels;
 
 import lombok.Data;
