+++ src/test/java/uk/gov/justice/laa/amend/claim/mappers/ClaimSummaryMapperTest.java
@@ -3,10 +3,10 @@ package uk.gov.justice.laa.amend.claim.mappers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
-import uk.gov.justice.laa.amend.claim.models.CivilClaimSummary;
-import uk.gov.justice.laa.amend.claim.models.ClaimFieldRow;
-import uk.gov.justice.laa.amend.claim.models.ClaimSummary;
-import uk.gov.justice.laa.amend.claim.models.CrimeClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.CivilClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.ClaimFieldRow;
+import uk.gov.justice.laa.amend.claim.viewmodels.ClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.CrimeClaimSummary;
 import uk.gov.justice.laa.dstew.payments.claimsdata.model.BoltOnPatch;
 import uk.gov.justice.laa.dstew.payments.claimsdata.model.ClaimResponse;
 import uk.gov.justice.laa.dstew.payments.claimsdata.model.FeeCalculationPatch;
