+++ src/test/java/uk/gov/justice/laa/amend/claim/views/ClaimSummaryViewTest.java
@@ -9,10 +9,10 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import uk.gov.justice.laa.amend.claim.config.LocalSecurityConfig;
 import uk.gov.justice.laa.amend.claim.controllers.ClaimSummaryController;
 import uk.gov.justice.laa.amend.claim.mappers.ClaimSummaryMapper;
-import uk.gov.justice.laa.amend.claim.models.CivilClaimSummary;
-import uk.gov.justice.laa.amend.claim.models.ClaimFieldRow;
-import uk.gov.justice.laa.amend.claim.models.ClaimSummary;
-import uk.gov.justice.laa.amend.claim.models.CrimeClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.CivilClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.ClaimFieldRow;
+import uk.gov.justice.laa.amend.claim.viewmodels.ClaimSummary;
+import uk.gov.justice.laa.amend.claim.viewmodels.CrimeClaimSummary;
 import uk.gov.justice.laa.amend.claim.service.ClaimService;
 import uk.gov.justice.laa.dstew.payments.claimsdata.model.ClaimResponse;
 import uk.gov.justice.laa.dstew.payments.claimsdata.model.FeeCalculationPatch;
